mesh render settings uses macros, shared with shaders

+++ include/vclib/render/mesh_render_settings_macros.h
@@ -27,33 +27,33 @@
  * These macros are used both on the library and on the shader side.
  */
 
-#define VCL_MRS_DRAW_MESH 1 << 0 // mesh visibility
+#define VCL_MRS_DRAW_MESH uint(1 << 0) // mesh visibility
 
 // points
-#define VCL_MRS_DRAW_POINTS              1 << 1 // point visibility
-#define VCL_MRS_DRAW_POINTS_PIXEL        1 << 2 // draw points as pixels
-#define VCL_MRS_DRAW_POINTS_CIRCLE       1 << 3 // draw points as circles
-#define VCL_MRS_DRAW_POINTS_COLOR_VERTEX 1 << 4 // use vertex colors for points
-#define VCL_MRS_DRAW_POINTS_COLOR_MESH   1 << 5 // use mesh color for points
-#define VCL_MRS_DRAW_POINTS_COLOR_USER   1 << 6 // use user color for points
+#define VCL_MRS_DRAW_POINTS              uint(1 << 1) // point visibility
+#define VCL_MRS_DRAW_POINTS_PIXEL        uint(1 << 2) // draw points as pixels
+#define VCL_MRS_DRAW_POINTS_CIRCLE       uint(1 << 3) // draw points as circles
+#define VCL_MRS_DRAW_POINTS_COLOR_VERTEX uint(1 << 4) // vert colors for points
+#define VCL_MRS_DRAW_POINTS_COLOR_MESH   uint(1 << 5) // mesh color for points
+#define VCL_MRS_DRAW_POINTS_COLOR_USER   uint(1 << 6) // user color for points
 
 // surface
-#define VCL_MRS_DRAW_SURF              1 << 10 // surface visibility
-#define VCL_MRS_DRAW_SURF_FLAT         1 << 11 // flat shading
-#define VCL_MRS_DRAW_SURF_SMOOTH       1 << 12 // smooth shading
-#define VCL_MRS_DRAW_SURF_COLOR_FACE   1 << 13 // use face color for surface
-#define VCL_MRS_DRAW_SURF_COLOR_VERTEX 1 << 14 // use vertex color for surface
-#define VCL_MRS_DRAW_SURF_COLOR_MESH   1 << 15 // use mesh color for surface
-#define VCL_MRS_DRAW_SURF_COLOR_USER   1 << 16 // use user color for surface
-#define VCL_MRS_DRAW_SURF_TEX_VERTEX   1 << 17 // use per vertex texcoords
-#define VCL_MRS_DRAW_SURF_TEX_WEDGE    1 << 18 // use per wedge texcoords
+#define VCL_MRS_DRAW_SURF              uint(1 << 10) // surface visibility
+#define VCL_MRS_DRAW_SURF_FLAT         uint(1 << 11) // flat shading
+#define VCL_MRS_DRAW_SURF_SMOOTH       uint(1 << 12) // smooth shading
+#define VCL_MRS_DRAW_SURF_COLOR_FACE   uint(1 << 13) // face color for surface
+#define VCL_MRS_DRAW_SURF_COLOR_VERTEX uint(1 << 14) // vertex color for surface
+#define VCL_MRS_DRAW_SURF_COLOR_MESH   uint(1 << 15) // mesh color for surface
+#define VCL_MRS_DRAW_SURF_COLOR_USER   uint(1 << 16) // user color for surface
+#define VCL_MRS_DRAW_SURF_TEX_VERTEX   uint(1 << 17) // per vertex texcoords
+#define VCL_MRS_DRAW_SURF_TEX_WEDGE    uint(1 << 18) // per wedge texcoords
 
 // wireframe
-#define VCL_MRS_DRAW_WIREFRAME            1 << 20 // draw wireframe
-#define VCL_MRS_DRAW_WIREFRAME_COLOR_USER 1 << 21 // user color for wireframe
-#define VCL_MRS_DRAW_WIREFRAME_COLOR_MESH 1 << 22 // mesh color for wireframe
+#define VCL_MRS_DRAW_WIREFRAME            uint(1 << 20) // draw wireframe
+#define VCL_MRS_DRAW_WIREFRAME_COLOR_USER uint(1 << 21) // user color for wrfrm
+#define VCL_MRS_DRAW_WIREFRAME_COLOR_MESH uint(1 << 22) // mesh color for wrfrm
 
 // bounding box
-#define VCL_MRS_DRAW_BOUNDINGBOX 1 << 25
+#define VCL_MRS_DRAW_BOUNDINGBOX uint(1 << 25) // bounding box visibility
 
 #endif // VCL_RENDER_MESH_RENDER_SETTINGS_MACROS_H
